Refactor driver's namespace
