+++ test/schema_validator/hyper_schema/router_test.rb
@@ -69,14 +69,14 @@ describe Committee::SchemaValidator::HyperSchema::Router do
   end
 
   def hyper_schema_router(options = {})
-    schema = Committee::Drivers::HyperSchema.new.parse(hyper_schema_data)
+    schema = Committee::Drivers::HyperSchema::Driver.new.parse(hyper_schema_data)
     validator_option = Committee::SchemaValidator::Option.new(options, schema, :hyper_schema)
 
     Committee::SchemaValidator::HyperSchema::Router.new(schema, validator_option)
   end
 
   def open_api_2_router(options = {})
-    schema = Committee::Drivers::OpenAPI2.new.parse(open_api_2_data)
+    schema = Committee::Drivers::OpenAPI2::Driver.new.parse(open_api_2_data)
     validator_option = Committee::SchemaValidator::Option.new(options, schema, :hyper_schema)
 
     Committee::SchemaValidator::HyperSchema::Router.new(schema, validator_option)
